ABAC working

+++ src/bvc_dec/bvc_dec.h
@@ -22,7 +22,7 @@ private:
 
 private:
 	uint8_t* output_picture_bytes;
-	uint32_t output_picture_size;
+	size_t	 output_picture_size;
 
 	std::shared_ptr<bvc_entropy_decoder> entropy_decoder;
 };
